mobile-version/server/requirements.txt
fastapi
uvicorn
sqlalchemy
pyyaml
pydantic
requests
imap-tools
python-telegram-bot
